Chatbot/requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-ollama
langchain-openai
langchain-huggingface
langchain_chroma
pypdf

#faiss-gpu # Biblioteca para armazenamento vetorial de documentos | faiss-gpu e faiss-cpu

# Ollama (cliente local)
ollama

# PyTorch (necessário para modelos locais com HuggingFace)
torch
# Nota: Para CUDA, instale a versão correta com: https://pytorch.org/get-started/locally/
# Ex: torch>=2.4.0+cu121 --index-url https://download.pytorch.org/whl/cu121

# Variáveis de ambiente
python-dotenv

# Opcional: para aceleração com transformers
transformers
accelerate